+++ src/Umbraco.Web.UI.Client/src/views/media/media.delete.controller.js
@@ -10,8 +10,12 @@ function MediaDeleteController($scope, mediaResource, treeService, navigationSer
 
     $scope.performDelete = function() {
 
+        // stop from firing again on double-click
+        if ($scope.busy) { return false; }
+
         //mark it for deletion (used in the UI)
         $scope.currentNode.loading = true;
+        $scope.busy = true;
 
         mediaResource.deleteById($scope.currentNode.id).then(function () {
             $scope.currentNode.loading = false;
@@ -26,6 +30,10 @@ function MediaDeleteController($scope, mediaResource, treeService, navigationSer
                 var recycleBin = treeService.getDescendantNode(rootNode, -21);
                 if (recycleBin) {
                     recycleBin.hasChildren = true;
+                    //reload the recycle bin if it's already expanded so the deleted item is shown
+                    if (recycleBin.expanded) {
+                        treeService.loadNodeChildren({ node: recycleBin, section: "media" });
+                    }
                 }
             }
             
@@ -33,18 +41,20 @@ function MediaDeleteController($scope, mediaResource, treeService, navigationSer
             if (editorState.current && editorState.current.id == $scope.currentNode.id) {
 
             	//If the deleted item lived at the root then just redirect back to the root, otherwise redirect to the item's parent
-            	var location = "/media";
-            	if ($scope.currentNode.parentId.toString() !== "-1")
+                var location = "/media";
+                if ($scope.currentNode.parentId.toString() === "-21")
+                    location = "/media/media/recyclebin";
+            	else if ($scope.currentNode.parentId.toString() !== "-1")
             		location = "/media/media/edit/" + $scope.currentNode.parentId;
 
                 $location.path(location);
             }
 
-            navigationService.hideMenu();
-
+            $scope.success = true;
         }, function (err) {
 
             $scope.currentNode.loading = false;
+            $scope.busy = false;
 
             //check if response is ysod
             if (err.status && err.status >= 500) {
